Improve server performance for data responses when streaming from the `data` child process (up to 2x)
Server process receives data by 8kb chunks. Previously received chunks are passed to a server response as is. That's a suboptimal for large data, e.g. for a 100Mb file it leads to ~12800 chunks which slowing down data transfering to a client. This is fixed by buffering received chunks from a child process to 1Mb, which reduces amount of chunks to send (for 100Mb it's 12800 -> 100) to a client and boosts data transfering up to 2 times as a result.